Confirmation to delete && and add a specific quantity

+++ src/pages/Admin/EditDish/index.jsx
@@ -19,10 +19,11 @@ export function EditDish() {
   const params = useParams();
 
   const [title, setTitle] = useState('');
-  const [costPrice, setCostPrice] = useState('');
-  const [salePrice, setSalePrice] = useState('');
+  const [costPrice, setCostPrice] = useState(0);
+  const [salePrice, setSalePrice] = useState(0);
   const [category, setCategory] = useState('');
-  const [quantity, setQuantity] = useState('');
+  const [quantity, setQuantity] = useState(0);
+  const [addQuantity, setAddQuantity] = useState(0);
 
   function handleCategory(e) {
     setCategory(e.target.value);
@@ -36,7 +37,7 @@ export function EditDish() {
       salePrice,
       costPrice,
       category,
-      quantity,
+      quantity: Number(quantity) + Number(addQuantity),
     });
 
     alert('Producto actualizado con éxito!');
@@ -45,6 +46,10 @@ export function EditDish() {
 
   async function handleDeleteProduct(e) {
     e.preventDefault();
+
+    if (!window.confirm('Estas seguro que deseas eliminar el producto?')) {
+      return;
+    }
 
     await api.delete(`/products/${params.id}`);
 
@@ -77,103 +82,124 @@ export function EditDish() {
           <Link to="/">
             <TextButton icon={BsChevronLeft} size={32} title="voltar" />
           </Link>
-          <h2>Actualizar producto</h2>
-          <Edit>
-            <LabelInput
-              type="text"
-              label="Nome"
-              placeholder="Ex.: Bomba Pila Gasolina"
-              gray
-              onChange={(e) => setTitle(e.target.value)}
-              value={title}
-            />
-            <label>
-              Categoria
-              <select id="category" onChange={handleCategory} value={category}>
-                <option value="">Selecione uma opção</option>
-                <option value="Bujias">Bujias</option>
-                <option value="">Selecione uma opção</option>
-                <option value="Bujias">Bujias</option>
-                <option value="Tren delantero">Tren delantero</option>
-                <option value="Terminal">Terminal</option>
-                <option value="Rotula">Rotula</option>
-                <option value="GEMELOS BARRA ESTABILIZADORA">
-                  GEMELOS BARRA ESTABILIZADORA
-                </option>
-                <option value="BUJE PEQUENO TIJERA DELANTERA">
-                  BUJE PEQUEÑO TIJERA DELANTERA
-                </option>
-                <option value="BUJE GRANDE TIJERA ">BUJE GRANDE TIJERA </option>
-                <option value="GOMAS U O GOMAS SECTOR">
-                  GOMAS U O GOMAS SECTOR
-                </option>
-                <option value="BASE AMORTIGUADOR DELANTERA">
-                  BASE AMORTIGUADOR DELANTERA
-                </option>
-                <option value="RODAMIENTO DELANTERO">
-                  RODAMIENTO DELANTERO
-                </option>
-                <option value="BOCINA DELANTERAS">BOCINA DELANTERAS</option>
-                <option value="MOZO TRASERO - RODAMIENTO TRASERO">
-                  MOZO TRASERO - RODAMIENTO TRASERO
-                </option>
-                <option value="RODAMIENTO BASE AMORTIGUADOR">
-                  RODAMIENTO BASE AMORTIGUADOR
-                </option>
-                <option value="GOMA BARRA ESTABILIZADORA DELANTERA">
-                  GOMA BARRA ESTABILIZADORA DELANTERA
-                </option>
-                <option value="TIJERA - BRAZOS OSCILANTES - DELANTERO">
-                  TIJERA - BRAZOS OSCILANTES - DELANTERO
-                </option>
-                <option value="PUNTAS DE TRIPOIDE">PUNTAS DE TRIPOIDE</option>
-                <option value="COPA CAJA">COPA CAJA</option>
-                <option value="TRICETA">TRICETA</option>
-                <option value="SOPORTES MOTOR Y CAJA">
-                  SOPORTES MOTOR Y CAJA
-                </option>
-                <option value="MUNON">MUNON</option>
-              </select>
-            </label>
-            <LabelInput
-              type="number"
-              label="Precio de Costo"
-              placeholder="$ 00,00"
-              gray
-              className="price"
-              onChange={(e) => setCostPrice(e.target.value)}
-              value={costPrice}
-            />
+          <section id="editView">
+            <h3>Actualizar producto</h3>
+            <Edit>
+              <div className="flex-item">
+                <LabelInput
+                  type="text"
+                  label="Nome"
+                  placeholder="Ex.: Bomba Pila Gasolina"
+                  gray
+                  onChange={(e) => setTitle(e.target.value)}
+                  value={title}
+                />
+                <label>
+                  Categoria
+                  <select
+                    id="category"
+                    onChange={handleCategory}
+                    value={category}
+                  >
+                    <option value="">Selecione uma opção</option>
+                    <option value="Bujias">Bujias</option>
+                    <option value="">Selecione uma opção</option>
+                    <option value="Bujias">Bujias</option>
+                    <option value="Tren delantero">Tren delantero</option>
+                    <option value="Terminal">Terminal</option>
+                    <option value="Rotula">Rotula</option>
+                    <option value="GEMELOS BARRA ESTABILIZADORA">
+                      GEMELOS BARRA ESTABILIZADORA
+                    </option>
+                    <option value="BUJE PEQUENO TIJERA DELANTERA">
+                      BUJE PEQUEÑO TIJERA DELANTERA
+                    </option>
+                    <option value="BUJE GRANDE TIJERA ">
+                      BUJE GRANDE TIJERA{' '}
+                    </option>
+                    <option value="GOMAS U O GOMAS SECTOR">
+                      GOMAS U O GOMAS SECTOR
+                    </option>
+                    <option value="BASE AMORTIGUADOR DELANTERA">
+                      BASE AMORTIGUADOR DELANTERA
+                    </option>
+                    <option value="RODAMIENTO DELANTERO">
+                      RODAMIENTO DELANTERO
+                    </option>
+                    <option value="BOCINA DELANTERAS">BOCINA DELANTERAS</option>
+                    <option value="MOZO TRASERO - RODAMIENTO TRASERO">
+                      MOZO TRASERO - RODAMIENTO TRASERO
+                    </option>
+                    <option value="RODAMIENTO BASE AMORTIGUADOR">
+                      RODAMIENTO BASE AMORTIGUADOR
+                    </option>
+                    <option value="GOMA BARRA ESTABILIZADORA DELANTERA">
+                      GOMA BARRA ESTABILIZADORA DELANTERA
+                    </option>
+                    <option value="TIJERA - BRAZOS OSCILANTES - DELANTERO">
+                      TIJERA - BRAZOS OSCILANTES - DELANTERO
+                    </option>
+                    <option value="PUNTAS DE TRIPOIDE">
+                      PUNTAS DE TRIPOIDE
+                    </option>
+                    <option value="COPA CAJA">COPA CAJA</option>
+                    <option value="TRICETA">TRICETA</option>
+                    <option value="SOPORTES MOTOR Y CAJA">
+                      SOPORTES MOTOR Y CAJA
+                    </option>
+                    <option value="MUNON">MUNON</option>
+                  </select>
+                </label>
+              </div>
+              <div className="flex-item">
+                <LabelInput
+                  type="number"
+                  label="Precio de Costo"
+                  placeholder="$ 00,00"
+                  gray
+                  className="price"
+                  onChange={(e) => setCostPrice(e.target.value)}
+                  value={costPrice}
+                />
 
-            <LabelInput
-              type="number"
-              label="Precio de Venta"
-              placeholder={!salePrice ? '$ 00,00' : `$ salePrice`}
-              gray
-              className="price"
-              onChange={(e) => setSalePrice(e.target.value)}
-              value={salePrice}
-            />
+                <LabelInput
+                  type="number"
+                  label="Precio de Venta"
+                  placeholder={!salePrice ? '$ 00,00' : `$ salePrice`}
+                  gray
+                  className="price"
+                  onChange={(e) => setSalePrice(e.target.value)}
+                  value={salePrice}
+                />
+                <LabelInput
+                  type="number"
+                  label="Cantidad"
+                  placeholder="Ex.: 10"
+                  gray
+                  className="quantity"
+                  value={quantity}
+                  onChange={(e) => setQuantity(e.target.value)}
+                />
+                <LabelInput
+                  type="number"
+                  label="Adicionar Cantidad"
+                  placeholder="Ex.: 10"
+                  gray
+                  className="quantity"
+                  onChange={(e) => setAddQuantity(e.target.value)}
+                />
+              </div>
 
-            <LabelInput
-              type="number"
-              label="Cantidad"
-              placeholder="Ex.: 10"
-              gray
-              className="quantity"
-              value={quantity}
-              onChange={(e) => setQuantity(e.target.value)}
-            />
-
-            <div className="options">
-              <button type="button" onClick={(e) => handleDeleteProduct(e)}>
-                Excluir prato
-              </button>
-              <button type="submit" onClick={(e) => handleUpdateProduct(e)}>
-                Salvar alterações
-              </button>
-            </div>
-          </Edit>
+              <div className="options">
+                <button type="button" onClick={(e) => handleDeleteProduct(e)}>
+                  Excluir prato
+                </button>
+                <button type="submit" onClick={(e) => handleUpdateProduct(e)}>
+                  Salvar alterações
+                </button>
+              </div>
+            </Edit>
+          </section>
         </Main>
       )}
